refactor: inline reward mapping helpers into tracking extension
Move RewardVerification mapping helpers next to their only call site and remove the now-redundant Present.swift file.
